+++ OurTicket/src/main/java/ko/ourticket/OurTicketApplication.java
@@ -1,4 +1,4 @@
-package Ko.OurTicket;
+package ko.ourticket;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
