+++ OurTicket/src/main/java/ko/ourticket/Ticket/GradeCount.java
@@ -1,4 +1,4 @@
-package Ko.OurTicket.Ticket;
+package ko.ourticket.Ticket;
 
 import java.util.Map;
