+++ OurTicket/src/main/java/ko/ourticket/Ticket/TicketController.java
@@ -1,7 +1,7 @@
-package Ko.OurTicket.performance;
+package ko.ourticket.Ticket;
 
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class PerformanceController {
+public class TicketController {
 }
