+++ OurTicket/src/main/java/ko/ourticket/performance/PerformanceController.java
@@ -1,7 +1,7 @@
-package Ko.OurTicket.Ticket;
+package ko.ourticket.performance;
 
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class TicketController {
+public class PerformanceController {
 }
