+++ OurTicket/src/main/java/ko/ourticket/performance/PerformanceRepository.java
@@ -1,4 +1,4 @@
-package Ko.OurTicket.performance;
+package ko.ourticket.performance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
