+++ OurTicket/src/test/java/ko/ourticket/OurTicketApplicationTests.java
@@ -1,4 +1,4 @@
-package Ko.OurTicket;
+package ko.ourticket;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
